interrupt support for CPU

+++ src/peripheral/mod.rs
@@ -6,6 +6,6 @@ pub mod interrupt_controller;
 /// - I/O (joypad, audio, graphics and LCD)
 /// - Interrupt controls.
 pub trait HardwareAccessible {
-    fn read_byte_from_hardware_register(&self, address: usize) -> u8;
-    fn write_byte_to_hardware_register(&mut self, address: usize, data: u8);
+    fn read_byte_from_hardware_register(&self, address: u16) -> u8;
+    fn write_byte_to_hardware_register(&mut self, address: u16, data: u8);
 }
